убрал из component методы task
