+++ src/task.js
@@ -33,6 +33,34 @@ export class Task extends Component {
     };
   }
 
+  get id() {
+    return this._id;
+  }
+
+  set id(id) {
+    this._id = id;
+  }
+
+  set id(id) {
+    this._id = id;
+  }
+
+  get dueDate() {
+    return this._dueDate;
+  }
+
+  get tags() {
+    return this._tags;
+  }
+
+  get color() {
+    return this._color;
+  }
+
+  get isRepeating() {
+    return this._isRepeating();
+  }
+
   render(getTemplates = this._state.editing ? getEditCardtemplate : getCardTemplate) {
     const newElement = document.createElement(`div`);
     const templateArguments = {
@@ -118,8 +146,27 @@ export class Task extends Component {
     return new Task(dataTask);
   }
 
-  static parseTasks(data) {
-    return data.map(Task.parseTask);
+  changeEditingStatus() {
+    this._state.editing = !this._state.editing;
+    this._state.isRepeat = this._isRepeating();
+  }
+
+  changeDateStatus() {
+    this._state.isDate = !this._state.isDate;
+  }
+
+  changeColor(color) {
+    this._color = color;
+  }
+
+  changeRepeatStatus() {
+    if (this._state.isRepeat) {
+      Object.keys(this._repeatingDays).forEach((day) => {
+        this._repeatingDays[day] = false;
+      });
+    }
+
+    this._state.isRepeat = !this._state.isRepeat;
   }
 
   toRAW() {
@@ -134,6 +181,14 @@ export class Task extends Component {
       'is_favorite': this.isFavorite,
       'is_done': this.isDone,
     };
+  }
+
+  _isRepeating() {
+    return Object.values(this._repeatingDays).some(Boolean);
+  }
+
+  static parseTasks(data) {
+    return data.map(Task.parseTask);
   }
 }
